backend/requirements.txt
# Django
# ------------------------------------------------------------------------------
django==3.1.5 # pyup: < 3.1  # https://www.djangoproject.com/
pytest-django==4.1.0  # https://github.com/pytest-dev/pytest-django
pytest==6.2.1  # https://github.com/pytest-dev/pytest
django-crispy-forms==1.10.0
social-auth-app-django ==4.0.0
Pillow==8.0.1
# ------------------------------------------------------------------------------

gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
